+++ app/src/main/java/ru/android_2019/citycam/cache/MyCache.java
@@ -1,12 +1,12 @@
 package ru.android_2019.citycam.cache;
 
 import android.util.LruCache;
-import ru.android_2019.citycam.CityCamActivity;
+import ru.android_2019.citycam.reader.WebcamsMessage;
 
 public class MyCache {
 
     private static MyCache instance;
-    private LruCache<String, CityCamActivity.DownloadImageTask.WebcamsMessage> lru;
+    private LruCache<String, WebcamsMessage> lru;
     private int cacheSize = 8 * 1024 * 1024; //8 MiB
 
     private MyCache() {
@@ -20,7 +20,7 @@ public class MyCache {
         return instance;
     }
 
-    public LruCache<String, CityCamActivity.DownloadImageTask.WebcamsMessage> getLru() {
+    public LruCache<String, WebcamsMessage> getLru() {
         return lru;
     }
 }
