+++ app/src/main/res/values/colors.xml
@@ -4,4 +4,7 @@
     <color name="colorPrimaryDark">#303F9F</color>
     <color name="colorAccent">#FF4081</color>
     <color name="colorAccentTransparent">#33FF4081</color>
+    <color name="colorCamBackground">#999</color>
+    <color name="colorBackground">#fff</color>
+    <color name="colorTable">#96ffc2</color>
 </resources>
