+++ app/src/main/res/values/dimens.xml
@@ -2,4 +2,6 @@
     <!-- Default screen margins, per the Android Design guidelines. -->
     <dimen name="activity_horizontal_margin">16dp</dimen>
     <dimen name="activity_vertical_margin">16dp</dimen>
+    <dimen name="table_column_width">160dp</dimen>
+    <dimen name="table_min_height">40dp</dimen>
 </resources>
